xrootd4j-gsi: refactor credential loading
Motivation:

A previous refactoring of the credential management
(ac1aa245f694161132848c552669806b5ce371fb, #11547)
inadvertently created unnecessary and inefficient
reinitialization/reloading of the credentials and
ca certs by moving that functionality out of the
factories and into the per-pipeline instance of
the GSICredentialManager.

Modification:

Factor out the credential loading from the manager
into a separate class to be injected by the
factories into the credential manager.

Also separate out loading from validation.

Result:

More efficient (from standpoint of cpu consumption)
loading and caching of certificates/credentials.

Target: master
Request: 3.5
Acked-by: Dmitry
